make ArchiveRecord a trait
